+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## [Unreleased]
 ### Added
 - [Issue #24](https://github.com/HardNorth/github-version-generate/issues/24) Data extraction, by @HardNorth
+- [Issue #31](https://github.com/HardNorth/github-version-generate/issues/31) Major, Minor, Patch, Prerelease, Buildmetadata fragments export, by @HardNorth
+
 
 ## [1.1.2]
 ### Changed